Group discretize and related functions
in a new discretize.py file.
