+++ nengo_loihi/hardware/builder.py
@@ -5,9 +5,9 @@ import logging
 
 import numpy as np
 
+from nengo_loihi.discretize import bias_to_manexp, vth_to_manexp
 from nengo_loihi.hardware.nxsdk_shim import (
     BasicSpikeGenerator, N2Board, TraceCfgGen)
-from nengo_loihi.model import bias_to_manexp, vth_to_manexp
 from nengo_loihi.synapses import tracing_mag_int_frac
 
 logger = logging.getLogger(__name__)
